mobileUserDeleteApi-LongId-22August2024 version 1

+++ MuseumTicketing/src/main/java/com/example/MuseumTicketing/appGuide/mainPara/qrCode/mobileReg/MobileRegRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MobileRegRepo extends JpaRepository<MobileReg,Integer> {
+public interface MobileRegRepo extends JpaRepository<MobileReg,Long> {
     Optional<MobileReg> findByPhNumber(String phNumber);
 }
